created separate MyInfo component and got it working properly with index.js
