+++ src/MyInfo.js
@@ -1,0 +1,17 @@
+import React from "react"
+
+function MyInfo() {
+  return (
+    <div>
+      <h1>Jacob Cavazos</h1>
+      <p>This is a paragraph about me...</p>
+      <ul>
+        <li>Japan</li>
+        <li>Turkey</li>
+        <li>Nordic Countries</li>
+      </ul>
+    </div>
+  )
+}
+
+export default MyInfo
