+++ tmp/dev-server/index.html
@@ -1,0 +1,1 @@
+<!DOCTYPE html><html lang="en-US" data-reactroot=""><head><meta charSet="UTF-8"/><meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=5, shrink-to-fit=no"/></head><body><div id="root"></div></body></html>
