+++ tmp/dev-server/react-static-browser-plugins.js
@@ -1,0 +1,12 @@
+// Imports
+
+
+// Plugins
+const plugins = [{
+  location: "/Users/jacobcavazos/React/scrimba-1",
+  plugins: [],
+  hooks: {}
+}]
+
+// Export em!
+export default plugins
