+++ tmp/dev-server/react-static-templates.js
@@ -1,0 +1,11 @@
+
+
+
+import t_0 from '../../node_modules/react-static/lib/browser/components/Default404'
+
+// Template Map
+export default {
+  '../../node_modules/react-static/lib/browser/components/Default404': t_0
+}
+
+export const notFoundTemplate = "../../node_modules/react-static/lib/browser/components/Default404"
